Alpha 1.0.5
Auto-identify and Auto-dust have been added. When enabled, these processes will attempt to replace the lowest rated crystal in the inventory with the newest crystal if the newest crystal is a higher rating than the lowest rated crystal. This means that autoidentifying millions of crystals searching for perfect crystals is now possible.

To add an extra layer of randomness to crystal rating, crystals can now be corrupted.  Corruption can only occur once on a crystal and offers positive or negative results.
Corruption has 4 potential results:
1.  Blessed- Tier increases by 1, rating and percent are adjusted accordingly
2. Cursed- Tier decreased by 1, rating and percent are adjusted accordingly
3. Destroyed- Crystal is destroyed.
4.  Plus- The first keyword gains a "+" modifier to it, this multiplies the rating of the crystal by 4.

This means that the current highest rating for a crystal is 52000. Ratings may be adjusted with the implementation of selling crystals, but for the time being this is the highest possible rating for a crystal, with the following keywords:
Perfect+ Illustrious Primal Ancient (52000)