Alpha 1.0.6
Auto-sell has been added. When enabled, it will grant the player money in return for crystals that are removed from the identified_crystals Array through the auto_identify_crystal method. This boolean cannot be enabled if autodust is already enabled, as the player would then be credited twice for a single crystal.

In addition, crystals can now be sorted using the sort_crystals method. This only works if there are no null values in the array, so a call to sort_crystals must be done before using the corrupt_all method.

Some GUI elements have appeared but these are merely placeholders while I become more comfortable with JPanel. Right now I am targeting the Alpha 1.10 release as the release of the first iteration of the GUI, but for now there is more that I want to add to the core game before adding in the GUI.